refactor(bot): break monitoring telemetry cycles (#871 PR 2)
Extracts singletons + classes from `SimplifiedTelemetry` into
`clients.ts` and shared interfaces into `telemetryTypes.ts`, so
`healthChecks.ts` and `telemetryMetrics.ts` no longer import back
through `SimplifiedTelemetry` to grab their dependencies.

Madge cycle count: 12 → 10 (Cluster B done; remaining are Cluster C
autoplay/queue + the type-only `types/CustomClient` cycle deferred to
follow-up PRs).

Public re-exports preserved on `SimplifiedTelemetry` for backwards
compatibility with `SimplifiedTelemetryWrapper`, `telemetry`,
`health`, `metrics`. No behavior changes.

Plan: .claude/plans/2026-05-16-bot-circular-deps.md (Phase 2)
ADR:  docs/decisions/2026-05-16-next-refactor-target-bot-circular-deps.md
